+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - New installations (not existing instances) now default to the `simple` full-text-search config
 - Unauthenticated search requests now by default force-disable remote fetches and pagination
 - Post search can now match text in the content warning with the database provider
+- prefixing a user search query with `@` will limit results to matching nicknames only, if the query contains no enclosed whitespace
 
 
 ## 2026.05 (3.19.0)